refactor: strict string amounts, remove float casts, type params, enum instances
- JournalLineData: tighten debit/credit from string|int|float to string
- Update all callers (ExpenseService, InvoiceService, FinalizeInvoiceAction, BankingService) to pass '0' string
- InvoiceService::yearlyRevenue: return string instead of float
- ExpenseQuery::yearlyTotal: return string instead of float
- ReportingService: remove (float) cast on bcsub net_profit
- DashboardService: use bcmath for balance, remove float casts on monetary values
- LedgerService: type computeTrialBalances parameter as Collection, fix stray closure bracket
- Expense::toSearchableArray: cast amount to string not float
- Pass enum instances directly (InvoiceStatus, ExpenseStatus) instead of ->value when model has enum cast
- Update test files to use string amounts

+++ app/Domains/Reporting/Services/ReportingService.php
@@ -37,7 +37,7 @@ class ReportingService
                 'expenses' => $expenses->values()->toArray(),
                 'total_revenue' => $totalRevenue,
                 'total_expenses' => $totalExpenses,
-                'net_profit' => (float) bcsub((string) $totalRevenue, (string) $totalExpenses, 2),
+                'net_profit' => bcsub((string) $totalRevenue, (string) $totalExpenses, 2),
             ];
         });
     }
